handle without any courses
